パーミッションに関するフォーマットを PermissionStyle に委譲
